Replace original `LatexSelect` with new `LatexSelect`
Add options to pass `sx` property for additional styling.
Fix component calls in `SWRReactionPickerImpl`, and
`EditForm`.
Remove superfluous dividers in `StateSelect` component.
