+++ app/src/shared/SWRStateSelect.tsx
@@ -13,7 +13,7 @@ import { Sx } from "@mantine/core";
 import { StateTree } from "@lxcat/database/dist/shared/queries/state";
 import { ReactionTypeTag } from "@lxcat/schema/dist/core/enumeration";
 import useSWRImmutable from "swr/immutable";
-import { StateSelect } from "./StateSelectNew";
+import { StateSelect } from "./StateSelect";
 
 export interface SWRReactionOptions {
   consumes: Array<StateLeaf>;
